requirements.txt
beautifulsoup4==4.3.2
boto==2.38.0
Django==1.7.11
django-appconf==1.0.1
django-compressor==1.4
django-libsass==0.2
django-modelcluster==0.6.2
django-storages-redux==1.3.1
django-taggit==0.12.3
django-treebeard==2.0
html5lib==0.999
libsass==0.7.0
Pillow==2.8.1
psycopg2==2.5.2
pytz==2015.2
requests==2.6.0
six==1.9.0
South==1.0
unicodecsv==0.12.0
Unidecode==0.4.17
wagtail==0.8.6
wheel==0.24.0
